+++ src/quick-lint-js/diag/diagnostic-types-2.h
@@ -191,6 +191,16 @@ struct Diag_Async_Export_Function {
   Source_Code_Span async_export;
 };
 
+struct Diag_Cyclic_TypeScript_Type_Definition {
+  [[qljs::diag("E0384", Diagnostic_Severity::error)]]  //
+  [[qljs::message("cannot use type directly in its own definition",
+                  ARG(use))]]                                            //
+  [[qljs::message("type {1} is being defined here", ARG(declaration))]]  //
+  Source_Code_Span use;
+  Source_Code_Span declaration;
+  Variable_Kind kind;
+};
+
 struct Diag_Declare_Class_Fields_Cannot_Have_Initializers {
   [[qljs::diag("E0335", Diagnostic_Severity::error)]]  //
   [[qljs::message("'declare class' fields cannot be initalized",